KAFKA-16123: move test cases to KStreamKStreamWindowCloseTest
